BreastCL_diagnosis/requirements.txt
numpy
pandas
scikit-learn
scipy
scikit-image
tqdm
PyYAML
torch
torchvision
monai
transformers
timm
fairscale
fairseq
open_clip_torch
einops
einops-exts
tutel
apex
